airflow/requirements.txt
apache-airflow-providers-slack>=8.0.0
apache-airflow-providers-snowflake>=6.0.0
dbt-core>=1.9.0
dbt-snowflake>=1.9.0
